+++ Model/Basket.php
@@ -21,6 +21,8 @@
 
 namespace OxidEsales\PayPalModule\Model;
 
+use OxidEsales\Eshop\Application\Model\Article as EshopArticle;
+
 /**
  * PayPal basket class
  *
@@ -258,5 +260,27 @@ class Basket extends Basket_parent
         }
 
         return $paymentVAT;
+    }
+
+    /**
+     * Check if variants of the given product are already in the basket
+     *
+     * @param EshopArticle $product
+     *
+     * @return bool
+     */
+    public function hasProductVariantInBasket(EshopArticle $product)
+    {
+        $return = false;
+
+        $variantIds = $product->getVariantIds();
+        foreach ($variantIds as $id) {
+            if ($this->getArtStockInBasket($id)) {
+                $return = true;
+                break;
+            }
+        }
+
+        return $return;
     }
 }